requirements.txt
fastapi
uvicorn
python-telegram-bot
aiohttp
requests
opencv-python-headless
pytesseract
pyzbar
Pillow
